requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
gitdb==4.0.9; python_version >= '3.6'
gitpython==3.1.27
smmap==5.0.0; python_version >= '3.6'
